% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_variance_explained}
\alias{pca_variance_explained}
\title{Get variance explained by each PC}
\usage{
pca_variance_explained(pca_result)
}
\arguments{
\item{pca_result}{A list containing a \code{prcomp} object in the
\code{pca} element, as returned by \code{\link{run_pca}}.}
}
\value{
A \code{data.frame} with columns:
\describe{
\item{PC}{Principal component name (e.g. \code{"PC1"}, \code{"PC2"})}
\item{variance_percent}{Percentage of total variance explained}
}
}
\description{
Calculates the percentage of total variance explained by each principal
component from a PCA result object.
}
\examples{
# res <- run_pca(example_se)
# var_exp <- pca_variance_explained(res)
# head(var_exp)

}
